WIP: Defines SparseMatrix_COO class and implementation
Description:
This commit adds the definitions for several members of the
SparseMatrix_COO class and adds some of their implementations in
SparseMatrix_COO.cpp. Therefore this commit does not build.
It adds definitions for:
 * private methods _disassemble and _getValueConcrete which
   must be defined on derived classes
 * COO datatype attributes these are implemented as raw pointers
 * Defines _fmt for MatrixFormat
 * Defines public methods assemble and matvec which must be
   defined on derived class

Not all routines are implemented. None are tested.
Commit should build, but code may have bugs.